commit com

+++ src/com/company/DAL/DAO.java
@@ -7,9 +7,6 @@ import java.util.List;
 public class DAO {
     private static Connection connection;
 
-    public static void closeConnection() throws SQLException {
-        connection.close();
-    }
     public static boolean openConection()  {
         try {
             Class.forName("org.sqlite.JDBC");
@@ -25,17 +22,27 @@ public class DAO {
         }
     }
     public static void dropTables() throws SQLException {
+        if (!openConection()) throw new SQLException();
+
         Statement query = connection.createStatement();
         query.execute("Drop table Clients");
         query.execute("Drop table Charters");
         query.execute("Drop table ProductAndCharterLinks");
         query.execute("Drop table Products");
+
+        query.close();
+        connection.close();
     }
     public static boolean createTableProducts(){
         try {
+            if (!openConection()) throw new SQLException();
+
             String text = "CREATE TABLE if not exists Products (title text primary key);";
             Statement query = connection.createStatement();
             query.executeUpdate(text);
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
@@ -44,9 +51,14 @@ public class DAO {
 
     public static boolean createTableClients(){
         try {
+            if (!openConection()) throw new SQLException();
+
             String text = "CREATE TABLE if not exists Clients (id INTEGER PRIMARY KEY , name text);";
             Statement query = connection.createStatement();
             query.executeUpdate(text);
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
@@ -54,10 +66,15 @@ public class DAO {
     }
     public static boolean createTableCharters(){
         try {
+            if (!openConection()) throw new SQLException();
+
             String text = "CREATE TABLE if not exists Charters (id INTEGER PRIMARY KEY, id_client INTEGER NOT NULL," +
                     " FOREIGN KEY(id_client) REFERENCES Clients(id) ON DELETE CASCADE);";
             Statement query = connection.createStatement();
             query.executeUpdate(text);
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
@@ -65,6 +82,8 @@ public class DAO {
     }
     public static boolean createTableProductAndCharterLinks(){
         try {
+            if (!openConection()) throw new SQLException();
+
             String text = "CREATE TABLE if not exists ProductAndCharterLinks (id INTEGER PRIMARY KEY," +
                     "title_product TEXT NOT NULL," +
                     "id_charter INTEGER NOT NULL," +
@@ -73,6 +92,8 @@ public class DAO {
             Statement query = connection.createStatement();
             query.executeUpdate(text);
 
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
@@ -81,10 +102,14 @@ public class DAO {
 
     public static boolean addProduct(String title){
         try{
+            if (!openConection()) throw new SQLException();
+
             String text = "INSERT into Products (title) VALUES ( + '" + title + "');";
             Statement query = connection.createStatement();
             query.executeUpdate(text);
 
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
@@ -92,25 +117,37 @@ public class DAO {
     }
     public static int addClient(String name){
         try{
+            if (!openConection()) throw new SQLException();
+
             String text = "INSERT into Clients (name) VALUES  ('" + name + "');";
             Statement query = connection.createStatement();
             query.executeUpdate(text);
             query = connection.createStatement();
-            ResultSet res = query.executeQuery("SELECT id FROM Clients WHERE id=last_insert_rowid()");
+            ResultSet result = query.executeQuery("SELECT id FROM Clients WHERE id=last_insert_rowid()");
+            int id = result.getInt("id");
 
-            return res.getInt("id");
+            query.close();
+            result.close();
+            connection.close();
+            return id;
         } catch (SQLException e) {
             return -1;
         }
     }
-    public static int addCharter(int idClient){
-        try{
+    public static int addCharter(int idClient) {
+        try {
+            if (!openConection()) throw new SQLException();
+
             String text = "INSERT into Charters (id_client) VALUES (" + idClient + ");";
             Statement query = connection.createStatement();
-            query.execute(text);
+            query.executeUpdate(text);
             query = connection.createStatement();
             ResultSet result = query.executeQuery("SELECT id FROM Charters WHERE id=last_insert_rowid();");
             int id = result.getInt("id");
+
+            query.close();
+            result.close();
+            connection.close();
             return id;
         } catch (SQLException e) {
             return -1;
@@ -118,9 +155,14 @@ public class DAO {
     }
     public static boolean addProductAndCharterLink(String titleProduct, int idCharter){
         try{
+            if (!openConection()) throw new SQLException();
+
             String text = "INSERT into ProductAndCharterLinks (title_product, id_charter) VALUES ( '" + titleProduct + "', " + idCharter + ");";
             Statement query = connection.createStatement();
-            query.execute(text);
+            query.executeUpdate(text);
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
@@ -128,6 +170,8 @@ public class DAO {
     }
 
     public static List<String> getProducts() throws SQLException {
+        if (!openConection()) throw new SQLException();
+
         List<String> result = new ArrayList<>();
         Statement query = connection.createStatement();
         ResultSet res = query.executeQuery("Select * from Products;");
@@ -135,9 +179,15 @@ public class DAO {
         while (res.next()){
             result.add(res.getString("title"));
         }
+
+        query.close();
+        res.close();
+        connection.close();
         return result;
     }
     public static List<String> getClients() throws SQLException {
+        if (!openConection()) throw new SQLException();
+
         List<String> result = new ArrayList<>();
         Statement query = connection.createStatement();
         ResultSet res = query.executeQuery("Select * from Clients;");
@@ -145,18 +195,29 @@ public class DAO {
             System.out.println("id: " + res.getInt("id") +" name: "+ res.getString("name"));
         }
 
+        query.close();
+        res.close();
+        connection.close();
         return result;
     }
     public static List<Integer> getCharters(int clientID) throws SQLException {
+        if (!openConection()) throw new SQLException();
+
         List<Integer> result = new ArrayList<>();
         Statement query = connection.createStatement();
         ResultSet res = query.executeQuery("Select id from Charters where id_client=" + clientID + ";");
         while (res.next()){
             result.add(res.getInt("id"));
         }
+
+        query.close();
+        res.close();
+        connection.close();
         return result;
     }
     public static List<String> getProductsInCharter(int id) throws SQLException {
+        if (!openConection()) throw new SQLException();
+
         List<String> result = new ArrayList<>();
         Statement query = connection.createStatement();
         String text = "Select title_product from ProductAndCharterLinks where id_charter=" + id + ";";
@@ -165,9 +226,14 @@ public class DAO {
            result.add(res.getString("title_product"));
         }
 
+        query.close();
+        res.close();
+        connection.close();
         return result;
     }
     public static List<Integer> getCharterOfProduct(String title) throws SQLException {
+        if (!openConection()) throw new SQLException();
+
         List<Integer> result = new ArrayList<>();
         Statement query = connection.createStatement();
         String text = "Select id_charter from ProductAndCharterLinks where title_product= '" + title + "';";
@@ -176,94 +242,127 @@ public class DAO {
             result.add(res.getInt("id_charter"));
         }
 
+        query.close();
+        res.close();
+        connection.close();
         return result;
     }
 
     public static boolean deleteClient(int clientID){
-        Statement query = null;
         try {
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             if (!deleteChartersOfClient(clientID)) throw new SQLDataException();
             query.executeUpdate("DELETE from Clients where id=" + clientID + ";");
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean deleteProduct(String title){
-        Statement query = null;
         try {
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             if (!deleteProductFromCharters(title)) throw new SQLDataException();
             query.executeUpdate("DELETE from Products where title='" + title + "';");
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean deleteCharter(int id){
-        Statement query = null;
         try {
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             if (!deleteCharterWithProducts(id)) throw new SQLDataException();
             query.executeUpdate("DELETE from Charters where id=" + id + ";");
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean deleteChartersOfClient(int idClient){
-        Statement query = null;
         try {
+
+
             List<Integer> charters = getCharters(idClient);
             for (int charter : charters){
                 if (!deleteCharter(charter)) throw new SQLDataException();
             }
+
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean deleteProductFromCharters(String title){
-        Statement query = null;
         try {
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             query.executeUpdate("DELETE from ProductAndCharterLinks where title_product='" + title + "';");
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean deleteCharterWithProducts(int id){
-        Statement query = null;
         try {
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             query.executeUpdate("DELETE from ProductAndCharterLinks where id_charter=" + id + ";");
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean deleteProductFromCharter(String title, int id_charter){
-        Statement query = null;
         try {
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             String text = "Select id from ProductAndCharterLinks limit 1";
             ResultSet res = query.executeQuery(text);
             System.out.println(res.getInt("id"));
             text = "DELETE from ProductAndCharterLinks where id_charter=" + id_charter + " and title_product='" + title + "' and id=" + res.getInt("id") + ";";
             query.executeUpdate(text);
+
+            query.close();
+            connection.close();
             return true;
         } catch (SQLException e) {
             return false;
         }
     }
     public static boolean updateClient(String name, int id){
-        Statement query = null;
         try{
-            query = connection.createStatement();
+            if (!openConection()) throw new SQLException();
+
+            Statement query = connection.createStatement();
             String text = "UPDATE Clients set name = '" + name + "' where id=" + id + ";";
             query.execute(text);
+
+            query.close();
+            connection.close();
             return true;
         }
         catch (Exception e){
